+++ example/lib/main-macos.dart
@@ -1,0 +1,37 @@
+// Copyright (c) 2023 Larry Aasen. All rights reserved.
+
+import 'package:flutter/material.dart';
+import 'package:upgrader/upgrader.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  // Only call clearSavedSettings() during testing to reset internal values.
+  await Upgrader.clearSavedSettings(); // REMOVE this for release builds
+
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final appcastURL =
+        'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast_macos.xml';
+    final cfg = AppcastConfiguration(url: appcastURL, supportedOS: ['macos']);
+
+    return MaterialApp(
+      title: 'Upgrader Example',
+      home: UpgradeAlert(
+          upgrader: Upgrader(
+            appcastConfig: cfg,
+            debugLogging: true,
+          ),
+          child: Scaffold(
+            appBar: AppBar(title: Text('Upgrader Example')),
+            body: Center(child: Text('Checking...')),
+          )),
+    );
+  }
+}